Tolerance Dictionary for df_comp function
In the previous iteration, my df_comp function had a list of columns for
comparison and a separate list of the associated tolerances. I decided
that would be better as a dictionary with keys being the columns and
values being the tolerances.
Made some changes to the IPython notebook as well to reflect these
changes.